admin jogosultság

+++ app/Http/Controllers/CategoryController.php
@@ -5,17 +5,42 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Category;
 use App\Http\Requests\CategoryRequest;
+use Illuminate\Support\Facades\Gate;
 
 
 class CategoryController extends Controller
 {
     public function getCategory(){
+        Gate::before(function($user){
+            $user = auth("sanctum")->user();
+            if($user -> role_id == 1){
+                return true;
+            }
+        }); 
+
+        if( !Gate::allows( "admin" )) {
+
+            return response()->json( ["Autentikációs hiba", "Nincs jogosultság"] );
+        }
+
         $categories = Category::all();
         return response()->json($categories);
     }
 
     public function addCategory(Request $request)
     {
+        Gate::before(function($user){
+            $user = auth("sanctum")->user();
+            if($user -> role_id == 1){
+                return true;
+            }
+        }); 
+
+        if( !Gate::allows( "admin" )) {
+
+            return response()->json( ["Autentikációs hiba", "Nincs jogosultság"] );
+        }
+
         $category = new Category();
         $category->name = $request["name"];
         $category->save();
@@ -25,12 +50,37 @@ class CategoryController extends Controller
 
     public function showCategory(Request $request)
     {
+
+        Gate::before(function($user){
+            $user = auth("sanctum")->user();
+            if($user -> role_id == 1){
+                return true;
+            }
+        }); 
+
+        if( !Gate::allows( "admin" )) {
+
+            return response()->json( ["Autentikációs hiba", "Nincs jogosultság"] );
+        }
         $category = Category::findOrFail($request["id"]);
         return response()->json($category);
     }
 
     public function updateCategory(Request $request)
     {
+
+        Gate::before(function($user){
+            $user = auth("sanctum")->user();
+            if($user -> role_id == 1){
+                return true;
+            }
+        }); 
+
+        if( !Gate::allows( "admin" )) {
+
+            return response()->json( ["Autentikációs hiba", "Nincs jogosultság"] );
+        }
+
         $category = Category::findOrFail($request["id"]);
         $category->name = $request["name"];
         $category->save();
@@ -40,6 +90,18 @@ class CategoryController extends Controller
 
     public function deleteCategory(Request $request)
     {
+        Gate::before(function($user){
+            $user = auth("sanctum")->user();
+            if($user -> role_id == 1){
+                return true;
+            }
+        }); 
+
+        if( !Gate::allows( "admin" )) {
+
+            return response()->json( ["Autentikációs hiba", "Nincs jogosultság"] );
+        }
+
         $category = Category::findOrFail($request["id"]);
         $category->delete();
